Initial commit of an in-development testing project called 'xpdeint', pronounced 'expedient'.
This code is written in python and generates C++ code. Currently it's very bare-bones, for example
the xml script the parser reads is hard-coded in the parser (parser2.py) and the simulation source
is sent to standard output. For build instructions, current and missing features, see the README
file in the xpdeint directory.

As a summary, the tools required to build and test xpdeint are:
Python 2.5 or later and Cheetah 2.0.1 or later.

xpdeint requires xmds to be installed for libxmds, and xmds_complex.h. Like xmds-1, if you want fourier
transforms, you need fftw2 or fftw3 installed, though at the moment, only fftw3 is supported.

Examples scripts are found in the examples subdirectory. These xmds scripts are using a new xmds-1 like
syntax, but because they are different, they require the xmds-version=2 attribute on the simulation element
so that they can be differentiated from scripts using the xmds-1 syntax.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1731 a0a4e76c-ba24-0410-a01c-b269113d594b