xpdeint: Gather information about which space a given vector is needed in.
This works by adding an additional 'dry run' stage where the template is 
converted to a string so that the necessary information can be gathered to be
made available the next time the template is converted to a string.

This will be useful so that a comprehensive set of plans can be generated at the
start of a simulation.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1790 a0a4e76c-ba24-0410-a01c-b269113d594b