xpdeint: Gather information about which space a given vector is needed in.
This works by adding an additional 'dry run' stage where the template is 
converted to a string so that the necessary information can be gathered to be
made available the next time the template is converted to a string.

This will be useful so that a comprehensive set of plans can be generated at the
start of a simulation.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1790 41b19578-b77a-4e3b-9682-c8570d237623