xpdeint: Fourier transforms (with FFTW3) are now integer-dimension safe.
As a side-effect, all fftw plans are created at the start of the simulation, even for mixed-space transforms. Using some of the work over the last couple of days, these plans contain all of the loops necessary for the transform. As a consequence, it is now faster to fourier transform just one dimension than it is to fourier transform all dimensions. Taking advantage of this, as diffusion.xmds only needs fourier transforms in one dimension for its IP operator, the simulation now runs 30% faster.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1792 41b19578-b77a-4e3b-9682-c8570d237623