xpdeint: Don't allocate a temporary vector for the purpose of planning when not strictly necessary.
In order to create fftw plans for vectors that haven't been allocated at plan creation time, a temporary vector was allocated equal to the size of the largest such vector that hadn't been allocated. This patch changes the strategy to use an already existing array if one is large enough. If there isn't one, then a temporary array is allocated as before.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1794 a0a4e76c-ba24-0410-a01c-b269113d594b