xpdeint: Add an example script for calculating the groundstate  wavefunction of a BEC
This is the first script to usefully use the filter operator. It has two filter operators, one to calculate the current normalisation of the wavefunction, and the second to correct it. The advantage of this groundstate script over the xmds-1 version is that it renormalises the wavefunction continuously, and in one integrate block, so adaptive integrators can be used.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1806 a0a4e76c-ba24-0410-a01c-b269113d594b