xpdeint: Edging closer with integer valued dimensions.
The parsing and replacement of EX operator strings in CDATA blocks has now been rewritten, and although you can't apply EX operators to arbitrary strings yet, the code is mostly there.

In the generated C code, components with integer valued dimensions (represented as phi[j*2, k, m-n][n, p % 3] in the XML) are mapped to
_phi(j*2, k, m-n, n, p % 3)
This is a better option than the other alternative which wouldn't have been
human-comprehensible.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1808 41b19578-b77a-4e3b-9682-c8570d237623