xpdeint: integer_valued_dimensions now work except for the delta a integrator in some cases.
The generated simulations compile, but will give incorrect results if the
delta a integrator code is of the form:
dx_dt = x[j-1]
The reason this fails is that currently, x[j-1] is overwritten with 
dx_dt[j-1]* _step on a previous loop iteration, so x[j-1]
isn't actually what you think it is. This only affects indices before the 'current'
index. i.e. dx_dt = x[j+1] would be safe.

However, you can do useful things with the current state of integer-valued
dimensions, kubo_integer_dimensions.xmds is an example. This simulation
runs all paths at once, and 'integration' over the integer valued dimension
can be used to calculate averages. The reason this example works and
integer_dimensions doesn't is that for the kubo example, all of the
dimensions are independent.

Also in this commit, Mattias' ArgvFeature has been reformatted.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1810 a0a4e76c-ba24-0410-a01c-b269113d594b