xpdeint: Integer-valued dimensions bugfixes and new example.
Added example gravity.xmds which simulations a system of
gravitationally-attracting particles.

_Integrator: Don't die if the sample count for a moment group is zero for the
integrator

_DeltaAOperator, DeltaAOperator: Initialise derivatives to zero for 
integer-valued dimensions.

StochasticFeature: Don't touch an integrator that isn't using noises.

parser2: Make the exit code of xpdeint the same as the exit code for the
compiler. This way, shell scripts can check for success or failure easily.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1820 41b19578-b77a-4e3b-9682-c8570d237623