xpdeint: Code cleanup and refactoring.
In the 'includes', 'defines', 'globals', etc functions there was a tension between placing @super at the start
(so that the superclass' implementation would occur first) and at the end (so that the code for the children
would appear after that of a given class' implementation). The solution was chosen differently for different
classes, and the whole implementation was bad. This has been fixed. Now just calling @super will only get the
code for the superclass, not for the children as well. The children can be called through calling 
implementationsForFunctionName('defines') for example. So it boils down to always call @super at the *start*

Various class structures have been made more sane. For example, the simulation drivers now inherit from
_ScriptElement not TopLevelSequenceElement. Code that didn't belong in TopLevelSequenceElement has been moved
to the simulation drivers.



git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1835 a0a4e76c-ba24-0410-a01c-b269113d594b