xpdeint: Optimisation for integer-valued-dimensions.
It is only necessary to initialise (and keep re-initialising) the created delta-a vectors
if the variables are being accessed out-of-order. gravity.xmds is a good example of this
where it calculates dvx_dt[k] += stuff as well as dvx_dt[j] += stuff at once. However,
when derivatives are not being accessed out-of-order, i.e. dpsi_dt[j], then there is no
need to reset the derivatives.

Also, minor cleanup of template instantiation in the template classes. They now use the same
trick as the parser (constructing an 'argumentsToTemplateConstructors' variable) instead of
using two variables which need to be explicitly passed to each template instantiation.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1836 41b19578-b77a-4e3b-9682-c8570d237623