xpdeint: Adding cool new examples.
wigner.xmds: Direct simulation of the evolution of the Wigner function for an
initial coherent state in an anharmonic oscillator (H = a^dagger a^dagger a a).
This simulation makes use of a filter operator during the integration to ensure that numerical artefacts at the edges do not arise. 

vibstring_*: These simulations use filter operators to implement arbitrary 
shape dirichlet boundary conditions. There is a simple 1D example of a wave
on a string being reflected at two edges, and two 2D examples. The first
is for waves with a circular boundary, and the second is for waves with an
elliptic boundary with the initial condition localised to one focus. In the elliptic
example, one can see part of the initial wave reconstructed at the opposite focus.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1844 a0a4e76c-ba24-0410-a01c-b269113d594b