xpdeint: Cross-propagation integrator now works.
Currently, only the RK4 integrator is able to be used as a cross-propagator.
This implementation of cross-propagation allows fixed-step integrators to be able to be used both as propagation
dimension integrators and cross-propagation integrators. The integrator has to say that it is capable of this though.
Currently only the RK4 is able to be used with the cross-propagator. In principle, the RK9 integrator could be made
to work as well, but that would require 9th order interpolation for any dependencies of the cross-propagation
integration, and that's a bit too much effort. However, a semi-implicit algorithm can (and should) be implemented
which would be able to be used both as a temporal integrator and as a cross-propagation integrator.

The code hasn't been checked with a 'right' cross-propagation boundary, but in principle, it should work.
Also, the value of the cross-propagation dimension is not available in the delta-a operator of the cross-propagation.
This will need to be fixed.

Other changes:
* If a template is a child of another template (i.e. it is in the 'children' attribute of another template), then you can
access the parent through the property 'parent'.
* super() is now called at the start of createNamedVectors/bindNamedVectors/preflight
* loopOverFieldInSpaceWithVectorsAndInnerContent now has an optional 'vectorsNotNeedingDefines' argument
which is a set of vectors for which the #define statements are not wanted.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1852 a0a4e76c-ba24-0410-a01c-b269113d594b