xpdeint: Documentation for the ScriptElement class.
Substantial overview documentation for the black-magic ScriptElement class.
Added a FIXME to FixedStepIntegrator about an issue with boundary conditions
that needs to be fixed.
Prevented the OutputFeature from adding code to cross-propagating integrators.

Oh, and I tested cross-propagation with a 'right' boundary, and it appears to work. So this
means that you can now have a simulation with *two* cross-propagation integrators going
in *opposite* directions! (Obviously the two integrators shouldn't depend on one another).

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1854 41b19578-b77a-4e3b-9682-c8570d237623