xpdeint: Documentation for the ScriptElement class.
Substantial overview documentation for the black-magic ScriptElement class.
Added a FIXME to FixedStepIntegrator about an issue with boundary conditions
that needs to be fixed.
Prevented the OutputFeature from adding code to cross-propagating integrators.

Oh, and I tested cross-propagation with a 'right' boundary, and it appears to work. So this
means that you can now have a simulation with *two* cross-propagation integrators going
in *opposite* directions! (Obviously the two integrators shouldn't depend on one another).

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1854 a0a4e76c-ba24-0410-a01c-b269113d594b