xpdeint: New noise sources using dSFMT (http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/SFMT/index.html)
dSFMT uses a Mersenne-Twister algorithm for generating the random numbers.
This is a higher-quality random number generator (in that it has better statistical
properties) than the linear congruential generator used by erand48 which is
what the POSIX noises are based on. In particular, the period of erand48
is 2^48, but the period of the Mersenne-Twister algorithm is 2^19937 - 1.

Another advantage is that dSFMT is faster than erand48. On taniwha, fibre.xmds
was 15% faster using a dSFMT-backed gaussian noise than using the erand48-backed
gaussian noise. On Mac OS X, this speedup is about 30%. The reason for the 
difference is that OS X's implementation of erand48 is simply slower than the one
in glibc.

To use the new DSFMT-backed noises, set the noise type to be 'gaussian-dsfmt',
'uniform-dsfmt' or 'poissonian-dsfmt'.

Currently the compiler options assume that the dSFMT.* files exist in the same
directory that ./parser2.py is called. Obviously we don't want this behaviour
when we release, but I don't know what the best solution is.

Other changes:
* The callOnceGuards have been moved to a new file and the implementation is
  a bit safer.
* The SlowGaussianPOSIXNoise has been removed. It doesn't really make sense
  to keep this around when it has no advantages over the faster version.




git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1860 41b19578-b77a-4e3b-9682-c8570d237623