xpdeint: Fix bug in automatic creation of the source field for Filter Operators.
The code would incorrectly add each character of a dimension name as
a name of a dimension that is needed in the source field. This works fine for
single-character dimension names, but fails for multiple-character dimension
names.

Also, use the dimensionWithName function instead of doing it manually. This
code was probably written before dimensionWithName existed.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1867 a0a4e76c-ba24-0410-a01c-b269113d594b