xpdeint: Fix bug in automatic creation of the source field for Filter Operators.
The code would incorrectly add each character of a dimension name as
a name of a dimension that is needed in the source field. This works fine for
single-character dimension names, but fails for multiple-character dimension
names.

Also, use the dimensionWithName function instead of doing it manually. This
code was probably written before dimensionWithName existed.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1867 41b19578-b77a-4e3b-9682-c8570d237623