xpdeint: Farewell filter operators, I knew thee well.
<operator kind="filter"> tags have been replaced with <filter> elements of
<filters> blocks inside <integrate> tags. These are now executed at the start
(or end) of an integration step, instead of every time the delta-a operator is
evaluated. This is much safer, as normalisation-changing behaviour isn't safe
inside a Runge-Kutta step.

<filters> tags can have a 'where' attribute with a value of either 'step start'
(default) or 'step end' which specifies when the filters in the <filters> tag get
executed.

Other changes:
* valueForKeyPath / setValueForKeyPath are now more helpful when an
  exception is hit. They will tell you what the keyPath was that they were trying
  to evaluate, and on what object.
* The way the arguments to operator evaluate() and calculate_operator_field()
  functions are handled is now much safer and more general.
* Examples have been updated to use the new notation.

TODO:
* <filter> tags cannot currently be used within <sample> blocks
* <operator> tags cannot currently be used within <sample> blocks 
  (it is unclear how to determine the field for an EX operator for a sampling
  block).

As a result of the complete removal of filter operators (and replacement with
either a computed_vector or a <filter> tag), the version number of xpdeint has
been updated to: '0.2 "Watch how easily we change notation..."'.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1871 41b19578-b77a-4e3b-9682-c8570d237623