xpdeint: Make the TODO file less scary.
'functions' (hacks) operators now exist. They can go before or after a CDATA
section in an operators tag to force the execution of the code before or after
the delta a operator. Note that 'functions' operators get executed in their own
function. You cannot rely on them existing in the same scope as the delta-a
block.

'filter' segments (xmds-1 filters) now exist. The notation is identical to the
notataion for our filter operators. So identical that it is parsed by the same
code, and a filter operator is what does the work for the filter segment.

Corrected a couple of bugs with moment groups with operators and 
computed vectors.

Add another example. This one calculates the groundstate of the
Schrödinger equation in an inverted gaussian potential.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1872 41b19578-b77a-4e3b-9682-c8570d237623