xpdeint: I forgot to run make check last night, as a result most things were broken.
Also, for a while it seems that stochastic simulations with fixed-step integrators and no error checking
were broken. I've modified kubo.xmds to now fit in this category so that we have a test for this.

Also, fixed 'make check' so it works after the directories got moved. Just a reminder, all examples should
compile except for vibstring3.xmds.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1904 41b19578-b77a-4e3b-9682-c8570d237623