xpdeint: I forgot to run make check last night, as a result most things were broken.
Also, for a while it seems that stochastic simulations with fixed-step integrators and no error checking
were broken. I've modified kubo.xmds to now fit in this category so that we have a test for this.

Also, fixed 'make check' so it works after the directories got moved. Just a reminder, all examples should
compile except for vibstring3.xmds.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1904 a0a4e76c-ba24-0410-a01c-b269113d594b