xpdeint: Argv elements can execute a block of code after argument processing.
I frequently found I needed to do post-processing on arguments for example
to construct a complex variable from real and imaginary parts, or to construct
an angular Rabi frequency from a 'bare' Rabi frequency. Previously this
required hacks. Now you just add a CDATA section to the end of your <argv>
element (by the way, does anyone think its a good idea to rename this to 
<args>?), and this code will get executed after arguments are processed.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1908 41b19578-b77a-4e3b-9682-c8570d237623