xpdeint: Fix up the indentation for some of the generated code.
The validation code wasn't being indented correctly because it was being
specified in the XMDS2Parser with additional whitespace. This didn't cause
any issues with the code except the extra indent was bugging me. So now the IndentFilter
supports the additional keyword argument 'fixupWhitespace'. If true, this will cause
the overall indent of a variable replacement to be cut to zero. Using this, user code
now has the correct indentation in the generated code and isn't indented by an extra
ten spaces or whatever it happened to be.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1912 a0a4e76c-ba24-0410-a01c-b269113d594b