xpdeint: No longer rely on xmds_complex.h
This commit creates an 'includes/' directory in the python egg distribution.
At compile-time we can work out the location of this directory, and add it to
the include path for the compiler.

Note that we still rely on libxmds in the XSIL loading code. So we still rely on
xmds-1.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1913 a0a4e76c-ba24-0410-a01c-b269113d594b