xpdeint: Computed Vectors can now access noises at evaluation.
These noises are calculated in the *looping field*. This only makes a
difference when a computed vector is calculated through an integral
over one or more dimensions. In these cases, the noises are evaluated
at each point in the larger space.

Noise vectors are now 'transform free' in that they will never cause
_ScriptElement's transformVectorsToSpace to throw an exception because
the noise vectors are 'double' vectors but are needed in a fourier-transformed
space.

Currently it isn't possible to turn off noises in computed vectors. To fix
this, we will need to add an 'evaluation' tag analogous to the 'initialisation'
tag for normal vectors. The CDATA code will need to go inside the evaluation
tag instead of just inside the computed_vector tag. Then you can specify which
noises are needed with a 'noises' attribute on the evaluation tag.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1932 41b19578-b77a-4e3b-9682-c8570d237623