xpdeint: Fix bug with noises and distributed-mpi.
Previously, if a noise was needed for a field that did not contain the distributed dimension,
then a different noise was generated on each node. This is not the desired behaviour. In this
case, the same noise should be used on each node. This patch causes the noise to only be calculated
on rank 0 and this noise is then distributed to the other nodes.

I've modified kubo_integer_dimensions_mpi.xmds to have a computed_vector that requires this. Note that
previous to this commit, kubo_integer_dimensions_mpi.xmds didn't actually use MPI. I forgot to turn it
on in the script...

Also, fixed a bug where noises would be generated in x-space even if the delta-a operator isn't in x-space.
I bet you didn't know you could have delta-a operators that aren't in x-space... You can also set the
'home space' for the integrator. This is the space in which the integration algorithm loops are performed.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1938 a0a4e76c-ba24-0410-a01c-b269113d594b