xpdeint: Minor fixes for the SI integrator.
Because SI calculates the derivative at the middle of the time-step,
the dependencies for cross-propagation are required at the mid-step.
This requires interpolation, but it doesn't need to be 4th order like
RK4, so Simpson's Rule interpolation is fine.

The delta-a operator space doesn't need to be the same as the home space
of the integrator, so a few more transformVectorsToSpace are needed. 


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1947 a0a4e76c-ba24-0410-a01c-b269113d594b