xpdeint: Add a 'simulation debugging' feature.
The idea behind this feature is that you turn it on by adding a '<simulation_debugging />' tag to the features list
and additional diagnostic information is printed as the simulation executes. The idea is that this should be useful
when a simulation is dying / giving bad results to help diagnose the cause. Currently, the output will tell you
the error of each component of a vector for adaptive timestep algorithms, and will print the attempted step size
for failed timesteps.

Perhaps this feature should be renamed to 'diagnostics'?


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@1962 a0a4e76c-ba24-0410-a01c-b269113d594b