xpdeint: Make filter operators take care of evaluating their own computed vectors.
There was an ordering issue with the excitedstate_gaussian.xmds example script where
after subtracting the projection of the current wavefunction with the groundstate, the
wavefunction was renormalised. But the normalisation vector was evaluated before the 
projection of the groundstate wavefunction was subtracted, hence the normalisation was
invalid. In the fix, all filter operators evaluate the computed vectors they need at the
start of the operator as another filter operator could have easily changed the outcome of
that computed vector.

This isn't necessary for other operators, as other operators cannot modify vectors directly.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1967 41b19578-b77a-4e3b-9682-c8570d237623