xpdeint: More progress towards distributed MPI for double-valued dimensions.
There is a potential bug with the coordinate variables when sampling in output moment groups
in fourier space with less than the full number of lattice points for that dimension. This needs
checking (added to TODO).

Anywhere where the size of a vector is needed, the templates now call on one of the following
functions as appropriate:
sizeOfVectorInSpace(vector, space): Return a string representing the local size of the vector in space.
sizeOfVectorInSpaceInReals(vector, space): Same as above except multiplied by 2 if the vector is complex.
sizeOfFieldInSpace(field, space): See the pattern?
allocSizeOfVector(vector): The number of points that should be allocated locally for this vector taking into
                           account swappedness/unswappedness.
allocSizeOfField(field): Similar to above.

Each of these functions call through to the simulation driver, and so they can be customised in the various
cases like distributed-over-integer-dimension, distributed-over-double, normal, etc.

As a result of these changes, the looping function 'loopOverVectorsWithInnerContent' had to change name to
'loopOverVectorsInSpaceWithInnerContent' because the space is required to determine the size of the field
(as it could be swapped or unswapped).

All simulations compile, but they need to be executed and the results checked.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@1991 41b19578-b77a-4e3b-9682-c8570d237623