xpdeint: Change the 'sd_' moment group prefix for multi-path simulations to 'stderr_'
For multi-path simulations, both the mean and the standard error of moment groups are output.
The mean in a variable of the form 'mean_momentVariableName' and in xmds-1 and in prior versions
of xpdeint, the standard error would be written in a variable of the form 'sd_momentVariableName'.
To avoid confusion with the standard deviation these variable names have now been renamed to 'stderr_'.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2005 a0a4e76c-ba24-0410-a01c-b269113d594b