xpdeint: The beginnings of a proper testing framework.
To run the tests, run './run_tests.py' in the xpdeint/ source directory.
This will then cause all of the simulations in the testsuite/ directory
to be compiled, executed and their results compared against known good results.

Note to use this, you must have the 'numpy' package installed. It comes installed
by default on Mac OS X Leopard, and will probably be installed by default with many
Python installs. If it isn't, just run 'easy_install numpy' or go to
http://numpy.scipy.org Note that numpy isn't needed to *use* xpdeint, only to run
the tests. And if you're modifying xpdeint, you should be running the tests before
every commit.

Features of this system:
* If the generated source for a given test hasn't changed since the last time the
  test passed on a given machine, the script isn't compiled or executed, saving
  time re-running tests that won't have changed.
* You can run a subset of the tests by typing './run_tests.py some_dir_name' where
  some_dir_name is a descendent of the testsuite directory. For example,
  './run_tests.py integrators'
* You can specify per-moment-group absolute and relative tolerances with which the
  results should be compared to the expected ones. This information is specified
  with a 'testing' tag in the xmds script. See the current examples in the testsuite/
  directory.
* Multiple tests can share the same expected results file.
* If the expected results file doesn't already exist, they will be generated from the
  output of executing the simulation the first time. This should only be needed when
  adding new tests as otherwise, the results will be stored in the repository.

Current limitations:
* Reading ASCII XSIL files has not yet been implemented, all output must be binary
  for the moment.
* Simulations that need to be executed with mpirun will currently be executed just
  by calling the simulation directly.
* There isn't a way to specify input files and have them copied across.

How it works:
For each testcase in turn, xpdeint generates the source file and puts it in a directory under
'/tmp/xpdeint/testsuite'. If this hasn't changed since the last time the test has passed
(using information stored in /tmp/xpdeint/testsuite), then the test passes. Note that version and
revision information is stripped from the generated source files so that there isn't a difference
in the generated source simply due to changing revision numbers. This source file is then compiled
and executed and the results (in /tmp/xpdeint/testsuite) compared with those stored in the repository.

This testsuite will replace the current 'make check' system which only compiles the examples and doesn't
check that the results they generate are sensible.

Other changes:
* Passing the option '-n' to xpdeint will prevent it from automatically compiling the generated source.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2022 41b19578-b77a-4e3b-9682-c8570d237623