xpdeint: Testing framework now supports MPI, supplying arguments to scripts and NaNs.
I've shifted some of the testscripts around and added a few more scripts. To add command-line arguments
to be used when testing a script, add the <arguments> tag inside the testing element. To change how
the script is called on the command line, add the <command_line> tag.
See testsuite/features/diffusion_arguments.xmds and testsuite/mpi/* for examples.



git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2025 a0a4e76c-ba24-0410-a01c-b269113d594b