xpdeint: Testing framework now supports MPI, supplying arguments to scripts and NaNs.
I've shifted some of the testscripts around and added a few more scripts. To add command-line arguments
to be used when testing a script, add the <arguments> tag inside the testing element. To change how
the script is called on the command line, add the <command_line> tag.
See testsuite/features/diffusion_arguments.xmds and testsuite/mpi/* for examples.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2025 41b19578-b77a-4e3b-9682-c8570d237623