xmds & xpdeint: CRITICAL BUG FIX to complex arithmetic.
The unary negation operator was incorrectly overloaded to modify the complex variable that was being negated.
Note that this bug only applies to the 'complex' class and not to the 'fftw_complex' class.

In more detail: Whenever code is written that requires the negative of a complex variable (of type complex, 
this bug is not present for the fftw_complex type), instead of simply returning the negative of that variable,
the variable itself was modified to be the negative as well. Consider the following code:

complex a = 7.0 + 3*i; // a = 7 + 3i, everything is fine
complex b = -a;
// b = -7 - 3i as expected, but additionally
// a = -7 - 3i

Note that this bug will not be triggered by a subtraction operation. b = a-c will have been correct.
Additionally, if the complex number is part of a larger expression, this bug will only matter if the
complex variable is first. i.e. z = -0.5*a would have been fine, but z = -a*0.5 would not. Note that
the variable 'i' was not affected by this because it is defined as 'const', so expressions like
z = -i*a were safe. Likewise any other 'const' complex constants would have also been safe.

From what I can gather from the repository logs, this bug has been present in xmds (and xpdeint) since the original
commit, i.e. February 3, 2003. This bug could have affected any simulation since then, though you would need
to be unlucky to hit it. Most simulations would be fine as the evolution equations are typically written in the form:
dphi_dt = L[phi] + stuff;
The L[phi] term will prevent the unary negation operation from being called. Unfortunately, some simulations will have hit
this. The only example script that hit this was the xmds van_der_pol.xmds script where its evolution equations are:
dx_dt = xdot;
dxdot_dt = -x - epsilon*(x*x - 1)*xdot;

The '-x' term in the second line triggered the bug, but the calculation of the x*x term is safe as although x is a complex
variable, its imaginary part is zero, and the memory that 'x' is stored in is immediately overwritten after this line with
the value of 'dx_dt * _step', and so although this simulation did hit the bug, due to a coincidence it did not affect the results.

This update should be considered critical and anyone following the development versions should update both xmds and xpdeint.

Joe: Can we do an xmds-1 release?


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2040 a0a4e76c-ba24-0410-a01c-b269113d594b