xpdeint: Auto-vectorisation of the cross-propagation operator is not safe.
Other changes:
 Minor changes to reduce compiler warnings.

It appears that my auto-vectorisation feature is playing badly with the cross-propagation
operator because there is a copy operation at the end of the step that works with 'complex'
types instead of the coerced 'double' types that makes gcc's auto-vectoriser happy. It appears
that this leads to the compiler not noticing the memory dependency between the final copy operation
and the last part of the integration step causing the compiler to happily reorder the writes so
that your integration step is essentially thrown away. See comments in _FixedStep.py for details.
The solution is to simply ban the auto-vectorisation feature from acting on cross-propagation operators.
Oh, and you won't hit this if you're using gcc-4.0. The only reason I hit it is I've been using gcc-4.2. Grr.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2042 41b19578-b77a-4e3b-9682-c8570d237623