xpdeint: Compiler errors are now reported using the line numbers in the source script!
Instead of compiler warnings/errors referring to the line number in the generated source file, this patch
causes them to refer to the line number in the .xmds script used to generate them making compiler
errors more meaningful to the user. For example, the following error:

vibstring.cc: In function ‘void _x_main_initialise()’:
vibstring.cc:622: error: ‘xy’ was not declared in this scope

will now read:
vibstring.xmds: In function ‘void _x_main_initialise()’:
vibstring.xmds:38: error: ‘xy’ was not declared in this scope

Note that any errors in lines that did not come directly from the source script will still refer
to the generated source file (you can tell this case as the filename will be end with '.cc').
If you want to turn this behaviour off because you need to find the line number in the generated
source file, just pass the '-d' or '--debug' option to xpdeint and the previous behaviour will be
used.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2045 41b19578-b77a-4e3b-9682-c8570d237623