xpdeint: Give delta-a cross-propagation operators access to their propagation dimension variable.
Previous to this patch, if you had a cross-propagation operator along the 'z' dimension then in your
cross-propagation integration code (the dphi_dz = stuff; bit) didn't have access to the 'z' variable.

This patch makes this possible by wrapping many functions in Function objects (a new class added in this
patch) which encapsulate the arguments that a function takes, its return type, etc. The C code to call
these functions is generated by calling 'call' on the Function object and passing in the arguments that
should be used when calling the function in the C code. Arguments can also be 'inherited' by a parent
Function object making it easy to pass arguments from one Function object to another.

This flexibility allows one object to modify the arguments taken by a function that belongs to another object
in a minimally intrusive way. Basically, the rest of the patch is adding functionality such that fixing
the actual bug can be done in about two lines.

A testcase has been added to test for this bug.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2050 a0a4e76c-ba24-0410-a01c-b269113d594b