xpdeint: Validation of input scripts!!!!
If you have ever misspelled a tag or an attribute and wondered why xmds/xpdeint didn't give you what you wanted,
then this commit is for you! If you have the python module lxml installed (sudo easy_install lxml) then assuming
xpdeint can read the script, it will then run the script through a RelaxNG schema (http://www.relaxng.org/)
and any violations will be printed as warnings. The schema is not designed to be incredibly strict, but it will
pick up any tag or attribute that xpdeint wouldn't read. Applying this schema to our examples and testcases did
pick up a few cases where old or invalid syntax was used.

The RelaxNG schema is in support/xpdeint.rnc in the compact syntax (i.e. the human-readable one). lxml doesn't
understand the compact syntax, only the XML syntax, so trang is used to convert between the two. To update the
support/xpdeint.rng file (the on in the XML syntax), just type 'make' in the support/ directory.

Note that installing lxml on Mac OS X is harder than it should be, if anyone wants to install it, send me an email,
and I'll send you a precompiled egg to use with easy_install.

If lxml isn't installed, then no validation will take place, but the scripts will still be compiled as per usual.



git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2064 a0a4e76c-ba24-0410-a01c-b269113d594b