xpdeint: The new SIC integrator now works.
The SIC integrator has cross-propagation combined in with the normal SI step
to make the combination more stable. This is somewhat of a reimplementation of
the algorithm used in xmds-1 when cross-propagation was used with the SIIP integrator.
Only now, you can have two cross-propagators going in opposite directions simultaneously!

Added testcases. Note that the boundary conditions specified for cross-propagation actually
mean something slightly different for the SIC integrator than they do for normal cross-propagation.
The boundary condition is applied at half a step outside the domain, instead of at the edge. This
is for numerical issues. This shouldn't cause a problem unless you are trying to construct tests
for the cross-propagator. See the new testcase for an example of how this influences things.

Vector's initialisation, goSpace and computed vector's evaluation functions now all use Function()
objects bringing all the flexibility that entails.

After discussing it with Andy, the loop index variable has been renamed again, this time to _index_${dimensionName}.
This only affects you if you're writing 'hacks'. I'm looking at you Michael.

Slight modification to the XML Schema for the cross-propagation operator using the SI integrator.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2080 41b19578-b77a-4e3b-9682-c8570d237623