xpdeint: Allow IP operators to be calculated as 'double's. (~15% Optimisation)
For typical quantum-mechanics problems, the IP operator is the kinetic energy operator,
k^2/2m... and this is a purely real quantity. We have up until this point assumed that
the IP operator is of type 'complex', which requires twice the storage of a 'double' and
more operators to multiply (and even more to divide). For this typical case, it is faster
to use a 'double' variable for the IP operator instead. I have seen speedups of ~14% for
both the diffusion and the groundstate_gaussian examples. To use this, just add the attribute
type='double' to your <operator kind="ip"> tag. For examples, see diffusion.xmds, groundstate_gaussian.xmds
or one of the new operator tests.

I have also added a number of tests for IP and EX operators as we had no explicit tests for those.
They all use the same expected results file so that we can ensure that they all give the same results.

Updated the XML Schema to reflect this addition.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2090 41b19578-b77a-4e3b-9682-c8570d237623