xpdeint: Bug fixes.
Allow EX operators to be used in moment groups again.
When diagnostics are turned on, print when steps succeed as well.
Change how the adaptive step size control works for adaptive integrators to try and make it more stable.
Correct a bug preventing IP operators working in scripts with integer-valued dimensions.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2105 41b19578-b77a-4e3b-9682-c8570d237623