xpdeint: Bug fixes.
Update testcases for the changes to the adaptive integrator made today.
I forgot to change the Stochastic feature after the change to the adaptive integrator (forgot to run the testsuite...).
Fixed a bug in the interaction between kind='first' integer dimensions, the distributed-mpi driver and fourier transforms.
Added a test 'anharmonic_oscillator_wigner_mpi' to check for this last bug.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2107 41b19578-b77a-4e3b-9682-c8570d237623