xpdeint: Bug fixes.
Update testcases for the changes to the adaptive integrator made today.
I forgot to change the Stochastic feature after the change to the adaptive integrator (forgot to run the testsuite...).
Fixed a bug in the interaction between kind='first' integer dimensions, the distributed-mpi driver and fourier transforms.
Added a test 'anharmonic_oscillator_wigner_mpi' to check for this last bug.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2107 a0a4e76c-ba24-0410-a01c-b269113d594b