xpdeint: Update TODO to add two new ideas.
The first is a general derivative notation that would be independent of the transform being used.
It would look something like D[phi, x, x, y, y] = d^4 phi/(dx^2 dy^2) (but I'm seeking better ideas for the syntax).

The second is the ability to separate IP operators when the derivative is separable (e.g. kx*kx + ky*ky). In the case
of adaptive step integrators in higher dimensions, by using this instead of normal IP operators, this should result in
significant savings of CPU time as the evaluation of the IP operator reduces from O(N^d) to O(d N) where d is the dimensionality
and N is the number of points of a given dimension. The reason for this is that in adaptive step integrators, IP operators require
the calculation of a complex exponential at each point at each timestep. This would reduce the number of points at which that complex
exponential needs to be evaluated.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2110 a0a4e76c-ba24-0410-a01c-b269113d594b