xpdeint: The distributed-mpi driver now supports fourier transforms again! Yay!
Bonus: It now supports distributed cosine and sine transforms too. The restriction is that
if the first dimension uses 'dft' transforms, then the second must, while if the first uses
either 'dct' or 'dst', the second can use *either* 'dct' or 'dst' transforms.

Added testcases, and we even pass the old MPI testcase that was used before I broke this feature.

Modified the "FFTW3MPI Design" files so that the content is clearer.

Additionally, this implementation of the distributed-mpi driver has no restrictions on subsampling dimensions.
The previous one refused to let you sub-sample the second dimension, while this version should support
arbitrary sub-sampling independent of the dimension, space or whether it's distributed or not.

All the testcases pass!


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2112 41b19578-b77a-4e3b-9682-c8570d237623